printer tests
